Remove generated code from PeopleController
Almost all changes made via prompt

>PeopleController was originally built by generating the code open api specs. We have since copied that code into our checked in code base and now generate the open api specs from code. I'd like to flatten PeopleController, removing the supertpye PeopleapiDelegate and related types PeopleApiController and PeopleApi, copying all required annotations onto PeopleController
>
>The controller should be tagged with `@Tag(name = "People”)` and any other tags defined on functions should be removed
